requirements.txt
streamlit
pandas
folium
streamlit-folium
altair
pillow
requests
beautifulsoup4
openai
openpyxl
python-docx
pypdf
